Reorganize Mersenne-31 crate for arch-specific implementations
